Natural-Language/Speech-Enabled-App/requirements.txt
python-dotenv~=1.0.0
azure-identity>=1.15.0
azure-cognitiveservices-speech==1.42.0